+++ backend-inventario/src/main/java/com/inventario/backend_inventario/Dto/AreaDto.java
@@ -2,6 +2,6 @@ package com.inventario.backend_inventario.Dto;
 import lombok.*;
 @Data
 public class AreaDto {
-    private Long id_area;
+    private Integer id_area;
     private String nombreArea;
 }
