+++ backend-inventario/src/main/java/com/inventario/backend_inventario/Dto/CategoriaDto.java
@@ -2,7 +2,7 @@ package com.inventario.backend_inventario.Dto;
 import lombok.*;
 @Data
 public class CategoriaDto {
-    private Long id_cat;
+    private Integer id_cat;
     private String nombreCat;
     private AreaDto area;
 }
