+++ backend-inventario/src/main/java/com/inventario/backend_inventario/Dto/ProveedorDto.java
@@ -2,6 +2,6 @@ package com.inventario.backend_inventario.Dto;
 import lombok.*;
 @Data   
 public class ProveedorDto {
-    private Long id_proveedor;
+    private Integer id_proveedor;
     private String nombre_proveedor;
 }
