+++ backend-inventario/src/main/java/com/inventario/backend_inventario/Repository/SedeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.inventario.backend_inventario.Model.Sede;
 
 @Repository
-public interface SedeRepository extends JpaRepository<Sede, Long> {
+public interface SedeRepository extends JpaRepository<Sede, Integer> {
     
 }
